fix: rename scene state hook
